refactor(firecracker): split firecracker_warm.rs god object into responsibility modules
firecracker_warm.rs (1667 LOC) -> firecracker_warm/ by responsibility:
config (knobs + host-memory reservation), host (WarmHost composition trait),
types (generation identity / claim / miss / outcome), serving (the seed /
refill / claim / retire engine). mod.rs keeps the module doc + shared imports
and re-exports the submodule surface; external API (WarmServing, WarmClaim,
TemplateIdentity, WarmLineage, configured_*, reserved_host_memory_mb, ...)
preserved via pub(crate) glob re-export. Tests relocated to
firecracker_warm/tests.rs. Pure code motion, no behavior change; max
production module now 434 LOC.
