+++ sandbox-runtime/src/firecracker_warm/tests.rs
@@ -1,806 +1,6 @@
-//! Firecracker warm-pool snapshot serving.
-//!
-//! Keeps `SANDBOX_FC_WARM_POOL_SIZE` *generations* warm. A generation is a
-//! self-contained unit of pre-provisioned identity:
-//!
-//! - a **template VM** (`fcwarm-g<N>-tpl`): cold-booted once on the
-//!   operator's configured base stack with its own TAP / vsock / rootfs,
-//!   paused, and snapshotted (`golden`). It is kept paused afterwards
-//!   because the snapshot artifacts live in its state dir — destroying it
-//!   would delete them (`microvm-runtime` layout).
-//! - a **rider TAP** (`fcwarm-g<N>-rider`): the host interface the pooled
-//!   entry restores onto via `SnapshotRef::network_overrides` (the template
-//!   still holds its own TAP while paused, so the entry cannot reuse it).
-//! - a [`WarmPool`] bucket of depth 1 holding one **pre-restored, paused
-//!   entry** ready for handoff.
-//!
-//! A claim is `acquire → rename_vm(entry, sandbox_id) → start_vm` — the
-//! handoff flow `microvm-runtime 0.4.0-alpha.2` implements `rename_vm` for.
-//! The claimed VM inherits the generation's identity: the template's guest
-//! IP (baked into the snapshot's memory image), the template's vsock UDS +
-//! CID (recorded in the vmstate), the template's rootfs backing file, and
-//! the rider TAP. Because that identity is single-occupancy, a generation
-//! serves exactly one claim: on claim the bucket is unregistered, the paused
-//! template is destroyed (releasing its memory and its TAP), and a fresh
-//! generation is seeded on the next create. Idle-time entry evictions
-//! (age / failed validation) do NOT retire the generation — the refill
-//! thread restores a replacement from the still-alive template snapshot,
-//! reusing the rider TAP sequentially.
-//!
-//! ## Admission-control invariant
-//!
-//! Pool inventory (templates + pre-restored entries) is NOT a live sandbox:
-//! it never enters the sandbox store. A warm claim itself is admitted exactly
-//! like a cold boot — the claim runs inside `firecracker::create_and_start`,
-//! which the runtime layer only calls AFTER `admit_sandbox_resources` +
-//! `enforce_sandbox_count_limit` have passed under the creation permit.
-//!
-//! The pool's own standing footprint is reserved against the host memory
-//! budget by [`reserved_host_memory_mb`]: `admit_sandbox_resources` adds
-//! `pool_size × 2 × mem_size_mib` (paused template + pre-restored entry per
-//! generation, the `file`-backend worst case) to committed memory, so
-//! `SANDBOX_HOST_MEMORY_BUDGET_MB` accounts for pool inventory even though it
-//! never enters the store. The factor stays 2 under `MICROVM_MEM_BACKEND=uffd`
-//! (a resumed guest can fault its entry pages fully in) to keep the guard
-//! fail-closed.
-//!
-//! ## Fail-loud boundaries
-//!
-//! The single designed fallback is warm-miss → cold boot, and every miss
-//! carries a typed [`WarmMiss`] reason that the create path logs. Seeding
-//! failures are logged loudly and retried on subsequent creates
-//! ([`WarmServing::ensure_seeding`]); a misconfigured
-//! `SANDBOX_FC_WARM_POOL_SIZE` is a hard error, never a silent disable.
-//!
-//! ## Restart reconciliation
-//!
-//! Pool VMs are process-local: after an operator restart the fresh process's
-//! provider maps are empty, so prior `fcwarm-*` templates and `warm-*` pool
-//! entries would keep running (holding guest memory) with no sandbox record.
-//! `firecracker::reconcile_warm_orphans` reaps them from `/proc` before the
-//! first seed — enumerating warm-prefixed Firecracker processes under the
-//! provider's socket dir, SIGKILLing them, and releasing their by-id host
-//! resources (TAP / vsock / rootfs clone / DNAT, plus the template's sibling
-//! rider TAP). It runs both as the first step of the warm-engine init and from
-//! `reaper::reconcile_on_startup` (covering the warm-disabled-now case). A
-//! `pkill -f 'fcwarm-'` alone is insufficient — it misses the `warm-*` pool
-//! entries that actually hold the restored guest memory.
+//! Test-support host + fixtures + the firecracker_warm unit tests.
 
-use std::net::Ipv4Addr;
-use std::path::PathBuf;
-use std::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
-
-use async_trait::async_trait;
-use microvm_runtime::{
-    model::{NetworkInterface, SnapshotRef, VmSpec, VmStatus},
-    provider::{VmProvider, VmQuery},
-};
-#[cfg(test)]
-use microvm_warm_pool::WarmPoolMetrics;
-use microvm_warm_pool::{EntryValidator, StackKey, ValidationResult, WarmPool, WarmPoolConfig};
-
-use crate::error::{Result, SandboxError};
-
-/// Snapshot name every generation's golden image is stored under.
-pub(crate) const GOLDEN_SNAPSHOT_ID: &str = "golden";
-
-/// Parse `SANDBOX_FC_WARM_POOL_SIZE`. Absent or `0` disables warm serving;
-/// anything unparseable is a hard configuration error — a typo must never
-/// silently disable the pool.
-pub(crate) fn configured_pool_size() -> Result<usize> {
-    match std::env::var("SANDBOX_FC_WARM_POOL_SIZE") {
-        Err(_) => Ok(0),
-        Ok(raw) => {
-            let trimmed = raw.trim();
-            if trimmed.is_empty() {
-                return Ok(0);
-            }
-            trimmed.parse::<usize>().map_err(|_| {
-                SandboxError::Validation(format!(
-                    "SANDBOX_FC_WARM_POOL_SIZE must be a non-negative integer, got {trimmed:?}"
-                ))
-            })
-        }
-    }
-}
-
-/// Paused template + pre-restored entry per generation — the `file`-backend
-/// worst case, kept for `uffd` too so the budget guard stays fail-closed.
-const WARM_GENERATION_MEM_FACTOR: u64 = 2;
-
-/// Standing host-memory footprint the warm pool reserves against
-/// `SANDBOX_HOST_MEMORY_BUDGET_MB`, so admission accounts for pool inventory
-/// that never enters the sandbox store. Zero when warm serving is disabled
-/// (`SANDBOX_FC_WARM_POOL_SIZE` unset/0), so Docker-only hosts are unaffected.
-///
-/// Deterministic from config — a fixed reservation, not a live pool query: the
-/// engine is lazily built only on the first Firecracker create (after
-/// admission), so a live query would report 0 reserved on the very create that
-/// is about to seed the pool and let it over-commit. `mem_size_mib` is read
-/// from the same `FirecrackerConfig::from_env()` the engine bakes into every
-/// generation's golden snapshot, so the reservation matches the real footprint.
-pub(crate) fn reserved_host_memory_mb() -> Result<u64> {
-    let pool_size = configured_pool_size()? as u64;
-    if pool_size == 0 {
-        return Ok(0);
-    }
-    let mem_size_mib =
-        microvm_runtime::adapters::firecracker::FirecrackerConfig::from_env().mem_size_mib as u64;
-    Ok(pool_size
-        .saturating_mul(WARM_GENERATION_MEM_FACTOR)
-        .saturating_mul(mem_size_mib))
-}
-
-/// Parse `SANDBOX_FC_WARM_MAX_AGE_SECS` (pool-entry age eviction). Default
-/// 3600s: evictions force a snapshot re-restore, so a long default keeps the
-/// pool quiet; operators lower it if they want fresher entries.
-pub(crate) fn configured_entry_max_age() -> Result<Duration> {
-    match std::env::var("SANDBOX_FC_WARM_MAX_AGE_SECS") {
-        Err(_) => Ok(Duration::from_secs(3600)),
-        Ok(raw) => raw
-            .trim()
-            .parse::<u64>()
-            .map(Duration::from_secs)
-            .map_err(|_| {
-                SandboxError::Validation(format!(
-                    "SANDBOX_FC_WARM_MAX_AGE_SECS must be a non-negative integer, got {raw:?}"
-                ))
-            }),
-    }
-}
-
-/// Parse `SANDBOX_FC_WARM_DISK_GB` — the per-generation rootfs clone size.
-/// `0` (default) keeps the provider's workspace-default rootfs (no clone),
-/// matching the cold path's `disk_gb == 0` semantics.
-pub(crate) fn configured_warm_disk_gb() -> Result<u64> {
-    match std::env::var("SANDBOX_FC_WARM_DISK_GB") {
-        Err(_) => Ok(0),
-        Ok(raw) => raw.trim().parse::<u64>().map_err(|_| {
-            SandboxError::Validation(format!(
-                "SANDBOX_FC_WARM_DISK_GB must be a non-negative integer, got {raw:?}"
-            ))
-        }),
-    }
-}
-
-/// Shape a warm generation is provisioned at and a create request must match
-/// (or leave unset) to be servable from the pool.
-#[derive(Debug, Clone)]
-pub(crate) struct WarmSettings {
-    /// Number of generations to keep claimable.
-    pub pool_size: usize,
-    /// Stack the template's rootfs is cloned from when `disk_gb > 0`
-    /// (`SANDBOX_FIRECRACKER_DEFAULT_STACK`). Irrelevant when `disk_gb == 0`
-    /// — the cold path ignores `image` in that case too, so warm/cold serve
-    /// identical guests.
-    pub stack: Option<String>,
-    /// Rootfs clone size for the template (`SANDBOX_FC_WARM_DISK_GB`).
-    pub disk_gb: u64,
-    /// vCPU count baked into the golden snapshot (provider workspace default).
-    pub vcpu_count: u8,
-    /// Memory baked into the golden snapshot (provider workspace default).
-    pub mem_size_mib: u32,
-    /// Pool-entry age eviction (`SANDBOX_FC_WARM_MAX_AGE_SECS`).
-    pub entry_max_age: Duration,
-}
-
-/// Host-side composition the engine cannot do itself (TAP / vsock / rootfs /
-/// guest readiness). Production wires the operator's `NetworkManager` /
-/// `VsockManager` / `RootfsRegistry` (see `firecracker::OperatorWarmHost`);
-/// tests and the network-less e2e substitute fakes.
-#[async_trait]
-pub(crate) trait WarmHost: Send + Sync + 'static {
-    /// Attach the template's own network + vsock + (optional) rootfs clone,
-    /// mutating `spec` in place. Returns the identity the generation's
-    /// eventual claimant inherits.
-    async fn compose_template(
-        &self,
-        template_id: &str,
-        disk_gb: u64,
-        stack: Option<&str>,
-        spec: &mut VmSpec,
-    ) -> Result<TemplateIdentity>;
-
-    /// Block until the guest inside the freshly booted template is ready to
-    /// be snapshotted (production: guest metadata daemon answers a ping).
-    async fn await_guest_ready(&self, template_id: &str, identity: &TemplateIdentity)
-    -> Result<()>;
-
-    /// Called once after the golden snapshot is written, before the bucket
-    /// is registered (i.e. before any restore can run). Production unlinks
-    /// the template's vsock UDS *file*: the paused template keeps an
-    /// orphaned listener fd (harmless — it never serves again), freeing the
-    /// path so the restored entry's Firecracker can bind the vsock device
-    /// recorded in the vmstate at the same path.
-    async fn prepare_snapshot_source(&self, template_id: &str, identity: &TemplateIdentity);
-
-    /// Attach the TAP the pooled entry restores onto. `None` when the host
-    /// composes no network (network-less e2e) — the snapshot then carries no
-    /// interface to override.
-    async fn attach_rider(&self, rider_id: &str) -> Result<Option<NetworkInterface>>;
-
-    /// Release template-keyed host resources. Flags select which: on claim,
-    /// only the TAP is released (vsock CID + rootfs clone move to the
-    /// claimed sandbox and are released at its delete); on seed failure,
-    /// everything is. Rider TAPs are never released through this trait:
-    /// `attach_rider` is the last fallible seeding step, and a claimed
-    /// rider is released at sandbox delete (`firecracker::release_attachments`).
-    async fn release_template(&self, template_id: &str, identity: &TemplateIdentity, all: bool);
-}
-
-/// Identity captured at template compose time, inherited by the claimant.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct TemplateIdentity {
-    /// Guest IP baked into the snapshot (kernel/network state). `None` only
-    /// for hosts that compose no network.
-    pub guest_ip: Option<Ipv4Addr>,
-    /// Host UDS path of the template's vsock — the claimed VM's metadata
-    /// listener binds here (recorded in the vmstate).
-    pub metadata_uds: Option<PathBuf>,
-    /// Whether a per-template rootfs clone was created (must be released
-    /// when the claimed sandbox is deleted).
-    pub rootfs_cloned: bool,
-}
-
-/// One claimable generation.
-#[derive(Debug, Clone)]
-pub(crate) struct Generation {
-    pub template_id: String,
-    pub rider_id: String,
-    pub stack_key: StackKey,
-    pub identity: TemplateIdentity,
-    pub rider_attached: bool,
-}
-
-/// Everything the create path needs to finish provisioning a claimed VM.
-#[derive(Debug, Clone)]
-pub(crate) struct WarmClaim {
-    /// Guest IP for the endpoint URL + DNAT (template's, via the snapshot).
-    pub guest_ip: Option<Ipv4Addr>,
-    /// UDS to inject per-sandbox env + sidecar token through.
-    pub metadata_uds: Option<PathBuf>,
-    /// Alias ids whose host resources now belong to the claimed sandbox.
-    pub lineage: WarmLineage,
-}
-
-/// Host resources a warm-claimed sandbox holds under ids other than its own.
-/// Released at sandbox delete in `firecracker::release_attachments`. Persisted
-/// by [`crate::firecracker_lineage`] keyed by sandbox id so a delete or
-/// reconcile after an operator restart can still release/reclaim them.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub(crate) struct WarmLineage {
-    /// Template id: owns the vsock CID allocation and (when
-    /// `rootfs_cloned`) the rootfs clone slot backing the claimed VM.
-    pub template_id: String,
-    /// Rider id: owns the TAP the claimed VM rides. `None` on network-less
-    /// hosts.
-    pub rider_id: Option<String>,
-    pub rootfs_cloned: bool,
-    /// Guest IP for endpoint rebuilds on resume (`firecracker::start`).
-    pub guest_ip: Option<Ipv4Addr>,
-}
-
-/// Typed warm-miss reason. Logged by the create path before the designed
-/// fallback to cold boot — the only silent thing about a miss is nothing.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum WarmMiss {
-    /// `SANDBOX_FC_WARM_POOL_SIZE` is 0/unset.
-    Disabled,
-    /// Pool enabled but no generation is ready yet (seeding in flight or
-    /// all seeds failed so far).
-    NotReady,
-    /// Ready generations exist but their buckets are empty (entry evicted
-    /// and not yet refilled, or refill failing).
-    Empty,
-    DiskMismatch {
-        requested: u64,
-        pooled: u64,
-    },
-    ImageMismatch {
-        requested: String,
-        pooled: String,
-    },
-    CpuMismatch {
-        requested: u64,
-        pooled: u8,
-    },
-    MemoryMismatch {
-        requested: u64,
-        pooled: u32,
-    },
-    /// Handoff rename failed; the entry was destroyed and the generation
-    /// returned to the pool for refill.
-    RenameFailed(String),
-    /// Post-rename resume failed; the claimed VM was destroyed and the
-    /// generation returned to the pool for refill.
-    ResumeFailed(String),
-}
-
-impl std::fmt::Display for WarmMiss {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            WarmMiss::Disabled => write!(f, "warm pool disabled (SANDBOX_FC_WARM_POOL_SIZE=0)"),
-            WarmMiss::NotReady => write!(f, "no warm generation ready (seeding in flight)"),
-            WarmMiss::Empty => write!(f, "warm buckets empty (entry evicted, refill pending)"),
-            WarmMiss::DiskMismatch { requested, pooled } => write!(
-                f,
-                "disk_gb mismatch (requested {requested}, pooled {pooled})"
-            ),
-            WarmMiss::ImageMismatch { requested, pooled } => {
-                write!(
-                    f,
-                    "image mismatch (requested {requested:?}, pooled {pooled:?})"
-                )
-            }
-            WarmMiss::CpuMismatch { requested, pooled } => {
-                write!(
-                    f,
-                    "cpu_cores mismatch (requested {requested}, pooled {pooled})"
-                )
-            }
-            WarmMiss::MemoryMismatch { requested, pooled } => {
-                write!(
-                    f,
-                    "memory_mb mismatch (requested {requested}, pooled {pooled})"
-                )
-            }
-            WarmMiss::RenameFailed(e) => write!(f, "warm handoff rename failed: {e}"),
-            WarmMiss::ResumeFailed(e) => write!(f, "warm handoff resume failed: {e}"),
-        }
-    }
-}
-
-/// Outcome of a warm-claim attempt.
-#[derive(Debug)]
-pub(crate) enum WarmOutcome {
-    Claimed(WarmClaim),
-    Miss(WarmMiss),
-}
-
-/// Request shape the claim gate evaluates (subset of
-/// `firecracker::FirecrackerCreateRequest` the engine needs — it must not
-/// see env/labels, which are injected after the claim).
-#[derive(Debug, Clone)]
-pub(crate) struct WarmClaimRequest {
-    pub sandbox_id: String,
-    pub image: String,
-    pub cpu_cores: u64,
-    pub memory_mb: u64,
-    pub disk_gb: u64,
-}
-
-/// Validates pool entries by provider record: present and paused
-/// (`Stopped`). Catches destroyed/renamed-away records; it cannot see a
-/// crashed FC child behind a live record — the claim path's resume would
-/// surface that as `ResumeFailed` (logged miss, cold fallback).
-struct PausedEntryValidator<P: VmQuery> {
-    provider: P,
-}
-
-impl<P: VmQuery + Send + Sync + 'static> EntryValidator for PausedEntryValidator<P> {
-    fn validate(&self, vm_id: &str) -> ValidationResult {
-        match self.provider.get_vm(vm_id) {
-            Ok(Some(view)) if view.status == VmStatus::Stopped => ValidationResult::Healthy,
-            Ok(Some(view)) => {
-                ValidationResult::Unhealthy(format!("entry {vm_id} in state {}", view.status))
-            }
-            Ok(None) => ValidationResult::Unhealthy(format!("entry {vm_id} has no record")),
-            Err(e) => ValidationResult::Unhealthy(format!("entry {vm_id} query failed: {e}")),
-        }
-    }
-}
-
-/// Seed/claim counters, exposed alongside [`WarmPoolMetrics`].
-#[derive(Debug, Default)]
-struct WarmCounters {
-    claims: AtomicU64,
-    misses: AtomicU64,
-    seed_failures: AtomicU64,
-}
-
-/// The warm-serving engine. Generic over the provider so tests drive it with
-/// an in-memory fake and the e2e with the real `FirecrackerVmProvider`.
-pub(crate) struct WarmServing<P: VmProvider + VmQuery + Clone + 'static> {
-    provider: P,
-    host: Arc<dyn WarmHost>,
-    settings: WarmSettings,
-    pool: WarmPool<P>,
-    ready: Mutex<Vec<Generation>>,
-    seeds_in_flight: Arc<AtomicUsize>,
-    generation_counter: AtomicU64,
-    counters: WarmCounters,
-}
-
-impl<P: VmProvider + VmQuery + Clone + 'static> WarmServing<P> {
-    pub(crate) fn new(provider: P, host: Arc<dyn WarmHost>, settings: WarmSettings) -> Self {
-        let pool = WarmPool::start(
-            provider.clone(),
-            WarmPoolConfig {
-                // Depth 1 per generation: a generation's identity (rider
-                // TAP, vsock UDS, rootfs file) is single-occupancy, so a
-                // deeper bucket would restore entries that collide on it.
-                min_depth: 1,
-                max_depth: 1,
-                refill_interval: Duration::from_secs(2),
-                entry_max_age: settings.entry_max_age,
-            },
-            Arc::new(PausedEntryValidator {
-                provider: provider.clone(),
-            }),
-        );
-        Self {
-            provider,
-            host,
-            settings,
-            pool,
-            ready: Mutex::new(Vec::new()),
-            seeds_in_flight: Arc::new(AtomicUsize::new(0)),
-            generation_counter: AtomicU64::new(0),
-            counters: WarmCounters::default(),
-        }
-    }
-
-    /// Number of generations currently claimable (bucket may still be
-    /// refilling; see [`WarmMiss::Empty`]).
-    pub(crate) fn ready_generations(&self) -> usize {
-        lock_ready(&self.ready).len()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn pool_metrics(&self) -> WarmPoolMetrics {
-        self.pool.metrics()
-    }
-
-    /// Top the pool back up to `pool_size` generations. Called by the create
-    /// path on every request: cheap when saturated, and it doubles as the
-    /// retry loop for failed seeds (no supervisor thread to babysit).
-    pub(crate) fn ensure_seeding(self: &Arc<Self>) {
-        loop {
-            let ready = self.ready_generations();
-            let in_flight = self.seeds_in_flight.load(Ordering::SeqCst);
-            if ready + in_flight >= self.settings.pool_size {
-                return;
-            }
-            // Reserve the slot before spawning so concurrent creates cannot
-            // over-seed past pool_size.
-            if self
-                .seeds_in_flight
-                .compare_exchange(in_flight, in_flight + 1, Ordering::SeqCst, Ordering::SeqCst)
-                .is_err()
-            {
-                continue;
-            }
-            let this = Arc::clone(self);
-            tokio::spawn(async move {
-                let outcome = this.seed_generation().await;
-                this.seeds_in_flight.fetch_sub(1, Ordering::SeqCst);
-                if let Err(err) = outcome {
-                    this.counters.seed_failures.fetch_add(1, Ordering::Relaxed);
-                    tracing::error!(%err, "firecracker warm-pool generation seed failed");
-                }
-            });
-        }
-    }
-
-    /// Cold-boot a template, snapshot it, and register the generation's
-    /// bucket. The pool's refill thread pre-restores the entry afterwards.
-    pub(crate) async fn seed_generation(&self) -> Result<()> {
-        let generation = self.generation_counter.fetch_add(1, Ordering::SeqCst);
-        let template_id = format!("fcwarm-g{generation}-tpl");
-        let rider_id = format!("fcwarm-g{generation}-rider");
-
-        let mut spec = VmSpec {
-            vcpu_count: Some(self.settings.vcpu_count),
-            mem_size_mib: Some(self.settings.mem_size_mib),
-            ..VmSpec::default()
-        };
-        let identity = self
-            .host
-            .compose_template(
-                &template_id,
-                self.settings.disk_gb,
-                self.settings.stack.as_deref(),
-                &mut spec,
-            )
-            .await?;
-
-        if let Err(err) = self
-            .boot_pause_snapshot(&template_id, &spec, &identity)
-            .await
-        {
-            // Roll back everything the template holds; the VM itself is
-            // destroyed (best-effort) inside boot_pause_snapshot.
-            self.host
-                .release_template(&template_id, &identity, true)
-                .await;
-            return Err(err);
-        }
-
-        self.host
-            .prepare_snapshot_source(&template_id, &identity)
-            .await;
-
-        let rider_iface = match self.host.attach_rider(&rider_id).await {
-            Ok(iface) => iface,
-            Err(err) => {
-                self.destroy_vm_best_effort(&template_id).await;
-                self.host
-                    .release_template(&template_id, &identity, true)
-                    .await;
-                return Err(err);
-            }
-        };
-        let rider_attached = rider_iface.is_some();
-
-        let stack_key = StackKey {
-            stack_name: self
-                .settings
-                .stack
-                .clone()
-                .unwrap_or_else(|| "workspace-default".to_string()),
-            version: format!("g{generation}"),
-            vcpu_count: self.settings.vcpu_count,
-            mem_size_mib: self.settings.mem_size_mib,
-        };
-        self.pool.register(
-            stack_key.clone(),
-            SnapshotRef {
-                vm_id: template_id.clone(),
-                snapshot_id: GOLDEN_SNAPSHOT_ID.to_string(),
-                // Entries wait paused; the claim resumes after rename so a
-                // pooled guest burns no CPU and its clock is corrected at
-                // handoff time, not pool-idle time.
-                resume_immediately: false,
-                network_overrides: rider_iface.into_iter().collect(),
-            },
-        );
-
-        lock_ready(&self.ready).push(Generation {
-            template_id: template_id.clone(),
-            rider_id,
-            stack_key,
-            identity,
-            rider_attached,
-        });
-        tracing::info!(
-            template_id,
-            generation,
-            "firecracker warm-pool generation seeded (golden snapshot registered)"
-        );
-        Ok(())
-    }
-
-    async fn boot_pause_snapshot(
-        &self,
-        template_id: &str,
-        spec: &VmSpec,
-        identity: &TemplateIdentity,
-    ) -> Result<()> {
-        let create_id = template_id.to_string();
-        let create_spec = spec.clone();
-        let provider = self.provider.clone();
-        run_provider(
-            move || provider.create_vm_with_spec(&create_id, &create_spec),
-            "warm template create",
-            template_id,
-        )
-        .await?;
-
-        let result: Result<()> = async {
-            let start_id = template_id.to_string();
-            let provider = self.provider.clone();
-            run_provider(
-                move || provider.start_vm(&start_id),
-                "warm template start",
-                template_id,
-            )
-            .await?;
-
-            self.host.await_guest_ready(template_id, identity).await?;
-
-            let pause_id = template_id.to_string();
-            let provider = self.provider.clone();
-            run_provider(
-                move || provider.stop_vm(&pause_id),
-                "warm template pause",
-                template_id,
-            )
-            .await?;
-
-            let snap_id = template_id.to_string();
-            let provider = self.provider.clone();
-            run_provider(
-                move || provider.snapshot_vm(&snap_id, GOLDEN_SNAPSHOT_ID),
-                "warm template snapshot",
-                template_id,
-            )
-            .await?;
-            Ok(())
-        }
-        .await;
-
-        if result.is_err() {
-            self.destroy_vm_best_effort(template_id).await;
-        }
-        result
-    }
-
-    /// Try to serve `request` from the pool. Never falls back internally —
-    /// the caller owns the (logged) cold-boot fallback.
-    pub(crate) async fn claim(&self, request: &WarmClaimRequest) -> WarmOutcome {
-        if let Some(miss) = self.shape_gate(request) {
-            return self.miss(miss);
-        }
-
-        // Pop a claimable generation whose bucket has an entry. The
-        // generation leaves `ready` before any await so a concurrent claim
-        // cannot double-serve it; on failure it is pushed back (its
-        // template + snapshot still exist, so the refill thread restores a
-        // replacement entry).
-        let (generation, handle) = {
-            let mut ready = lock_ready(&self.ready);
-            let mut found = None;
-            for (idx, generation) in ready.iter().enumerate() {
-                if let Some(handle) = self.pool.acquire(&generation.stack_key) {
-                    found = Some((idx, handle));
-                    break;
-                }
-            }
-            match found {
-                Some((idx, handle)) => (ready.remove(idx), handle),
-                None => {
-                    drop(ready);
-                    let reason = if self.ready_generations() == 0 {
-                        WarmMiss::NotReady
-                    } else {
-                        WarmMiss::Empty
-                    };
-                    return self.miss(reason);
-                }
-            }
-        };
-
-        let entry_id = handle.source_vm_id.clone();
-        let sandbox_id = request.sandbox_id.clone();
-
-        // Handoff: re-key the pre-restored entry onto the sandbox id.
-        let rename_provider = self.provider.clone();
-        let rename_entry = entry_id.clone();
-        let rename_target = sandbox_id.clone();
-        if let Err(err) = run_provider(
-            move || rename_provider.rename_vm(&rename_entry, &rename_target),
-            "warm handoff rename",
-            &entry_id,
-        )
-        .await
-        {
-            self.destroy_vm_best_effort(&entry_id).await;
-            lock_ready(&self.ready).push(generation);
-            return self.miss(WarmMiss::RenameFailed(err.to_string()));
-        }
-
-        // Entries are restored paused; resume under the new identity.
-        let resume_provider = self.provider.clone();
-        let resume_id = sandbox_id.clone();
-        if let Err(err) = run_provider(
-            move || resume_provider.start_vm(&resume_id),
-            "warm handoff resume",
-            &sandbox_id,
-        )
-        .await
-        {
-            self.destroy_vm_best_effort(&sandbox_id).await;
-            lock_ready(&self.ready).push(generation);
-            return self.miss(WarmMiss::ResumeFailed(err.to_string()));
-        }
-
-        // The generation is spent: its identity now belongs to the claimed
-        // sandbox. Unregister the bucket (a refill racing this window would
-        // restore an entry the unregister immediately destroys — upstream
-        // WarmPool destroys orphans inserted into unregistered buckets),
-        // destroy the paused template (frees its guest memory and TAP; the
-        // snapshot artifacts die with it, which is fine — nothing restores
-        // from a spent generation), and keep the vsock CID + rootfs clone
-        // alive under the template id until the sandbox is deleted.
-        self.pool.unregister(&generation.stack_key);
-        self.destroy_vm_best_effort(&generation.template_id).await;
-        self.host
-            .release_template(&generation.template_id, &generation.identity, false)
-            .await;
-
-        self.counters.claims.fetch_add(1, Ordering::Relaxed);
-        tracing::info!(
-            sandbox_id = %request.sandbox_id,
-            template_id = %generation.template_id,
-            "firecracker warm-pool claim served (snapshot-restored VM handed off via rename)"
-        );
-
-        WarmOutcome::Claimed(WarmClaim {
-            guest_ip: generation.identity.guest_ip,
-            metadata_uds: generation.identity.metadata_uds.clone(),
-            lineage: WarmLineage {
-                template_id: generation.template_id,
-                rider_id: generation.rider_attached.then_some(generation.rider_id),
-                rootfs_cloned: generation.identity.rootfs_cloned,
-                guest_ip: generation.identity.guest_ip,
-            },
-        })
-    }
-
-    /// `None` = request matches the pooled shape.
-    fn shape_gate(&self, request: &WarmClaimRequest) -> Option<WarmMiss> {
-        if request.disk_gb != self.settings.disk_gb {
-            return Some(WarmMiss::DiskMismatch {
-                requested: request.disk_gb,
-                pooled: self.settings.disk_gb,
-            });
-        }
-        // When disk_gb == 0 the cold path boots the workspace-default rootfs
-        // regardless of `image`, so warm serves any image for parity. With a
-        // cloned template the image must name the pooled stack (or be empty
-        // and default to it).
-        if self.settings.disk_gb > 0 {
-            let pooled = self.settings.stack.clone().unwrap_or_default();
-            let requested = request.image.trim();
-            if !requested.is_empty() && requested != pooled {
-                return Some(WarmMiss::ImageMismatch {
-                    requested: requested.to_string(),
-                    pooled,
-                });
-            }
-        }
-        if request.cpu_cores != 0 && request.cpu_cores != u64::from(self.settings.vcpu_count) {
-            return Some(WarmMiss::CpuMismatch {
-                requested: request.cpu_cores,
-                pooled: self.settings.vcpu_count,
-            });
-        }
-        if request.memory_mb != 0 && request.memory_mb != u64::from(self.settings.mem_size_mib) {
-            return Some(WarmMiss::MemoryMismatch {
-                requested: request.memory_mb,
-                pooled: self.settings.mem_size_mib,
-            });
-        }
-        None
-    }
-
-    fn miss(&self, miss: WarmMiss) -> WarmOutcome {
-        self.counters.misses.fetch_add(1, Ordering::Relaxed);
-        WarmOutcome::Miss(miss)
-    }
-
-    async fn destroy_vm_best_effort(&self, vm_id: &str) {
-        let provider = self.provider.clone();
-        let id = vm_id.to_string();
-        let owner = vm_id.to_string();
-        if let Err(err) =
-            run_provider(move || provider.destroy_vm(&id), "warm destroy", &owner).await
-        {
-            tracing::warn!(vm_id = %owner, %err, "failed destroying warm-pool VM");
-        }
-    }
-}
-
-/// Provider calls shell out / do unix-socket I/O; keep them off the async
-/// runtime. Mirrors `firecracker.rs`'s spawn_blocking convention.
-async fn run_provider<F>(f: F, action: &str, vm_id: &str) -> Result<()>
-where
-    F: FnOnce() -> std::result::Result<(), microvm_runtime::error::VmRuntimeError> + Send + 'static,
-{
-    tokio::task::spawn_blocking(f)
-        .await
-        .map_err(|e| SandboxError::Unavailable(format!("{action} join error for {vm_id}: {e}")))?
-        .map_err(|e| SandboxError::Unavailable(format!("{action} for {vm_id}: {e}")))
-}
-
-fn lock_ready(m: &Mutex<Vec<Generation>>) -> std::sync::MutexGuard<'_, Vec<Generation>> {
-    match m.lock() {
-        Ok(guard) => guard,
-        Err(poisoned) => poisoned.into_inner(),
-    }
-}
+use super::*;
 
 /// Test / e2e host that composes no network, no vsock, and no rootfs clone.
 /// The engine's lifecycle (seed → refill → claim → retire) is identical;
@@ -859,7 +59,7 @@ pub(crate) fn test_settings(pool_size: usize) -> WarmSettings {
 }
 
 #[cfg(test)]
-mod tests {
+mod warm_serving {
     use super::*;
     use std::collections::HashMap as StdHashMap;
     use std::time::Instant;
