+++ scripts/.god-object-baseline.txt
@@ -2,10 +2,9 @@
 # Regenerate after a real split: scripts/check-file-sizes.sh --update
 # Doctrine: CLAUDE.md 'Module Size Discipline' — split by responsibility toward <600 LOC.
 2966 ui/src/routes/create.tsx
-1631 sandbox-runtime/src/firecracker_warm.rs
 1483 sandbox-runtime/src/tee/mod.rs
+1423 sandbox-runtime/src/firecracker.rs
 1281 ui/src/routes/sandboxes.$id.tsx
-1156 sandbox-runtime/src/firecracker.rs
 1111 ai-agent-sandbox-blueprint-bin/src/main.rs
 1106 ui/src/routes/instances.$id.tsx
 1095 ai-agent-sandbox-blueprint-lib/src/workflows.rs
@@ -13,7 +12,7 @@
 948 ai-agent-instance-blueprint-lib/src/workflows.rs
 922 ui/src/routes/workflows._index.tsx
 884 sandbox-runtime/src/tee/verify.rs
-831 sandbox-runtime/src/reaper.rs
+838 sandbox-runtime/src/reaper.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
 721 ui/src/components/shared/SessionSidebar.tsx
 701 ui/src/lib/stores/chatSessions.ts
